# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: IAM Service API
  version: 0.1.0
  description: |
    Аутентификация, организации, членство и RBAC. Префикс `/v1` может навешиваться API Gateway.
servers:
  - url: https://api.example.com/v1/iam
    description: через gateway
  - url: http://localhost:8080/v1/iam
    description: локально
...
security:
  - bearerAuth: []
paths:
...
  /organizations:
    post:
      tags: [organizations]
      summary: Создать организацию (создатель — owner)
      operationId: orgCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationCreateRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '403':
          $ref: '#/components/responses/Forbidden'
  /organizations/{orgId}:
    get:
      tags: [organizations]
      summary: Карточка организации
      operationId: orgGet
      parameters:
        - $ref: '#/components/parameters/OrgId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [organizations]
      summary: Обновить организацию
      operationId: orgPatch
      parameters:
        - $ref: '#/components/parameters/OrgId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationPatchRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '403':
          $ref: '#/components/responses/Forbidden'
  /organizations/{orgId}/members:
    get:
      tags: [organizations]
      summary: Список членов
      operationId: orgMembersList
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberPage'
  /organizations/{orgId}/members/{membershipId}:
    patch:
      tags: [organizations]
      summary: Смена ролей / статуса членства
      operationId: orgMemberPatch
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/MembershipId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MembershipPatchRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Membership'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags: [organizations]
      summary: Исключить члена
      operationId: orgMemberDelete
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/MembershipId'
      responses:
        '204':
          description: OK
        '403':
          $ref: '#/components/responses/Forbidden'
  /organizations/{orgId}/invites:
    get:
      tags: [organizations]
      summary: Список инвайтов
      operationId: orgInvitesList
      parameters:
        - $ref: '#/components/parameters/OrgId'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Invite'
    post:
      tags: [organizations]
      summary: Создать инвайт
      operationId: orgInvitesCreate
      parameters:
        - $ref: '#/components/parameters/OrgId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteCreateRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteCreated'
  /organizations/{orgId}/invites/{inviteId}:
    delete:
      tags: [organizations]
      summary: Отозвать инвайт
      operationId: orgInvitesRevoke
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - name: inviteId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: OK
...
  /organizations/{orgId}/idp:
    get:
      tags: [idp]
      summary: Настройки IdP (маскированные секреты)
      operationId: orgIdpGet
      parameters:
        - $ref: '#/components/parameters/OrgId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgIdpConfigPublic'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [idp]
      summary: Обновить IdP
      operationId: orgIdpPatch
      parameters:
        - $ref: '#/components/parameters/OrgId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrgIdpConfigPatchRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgIdpConfigPublic'
  /organizations/{orgId}/idp/test:
    post:
      tags: [idp]
      summary: Проверка metadata / discovery
      operationId: orgIdpTest
      parameters:
        - $ref: '#/components/parameters/OrgId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrgIdpTestRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgIdpTestResult'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    MembershipId:
      name: membershipId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    Forbidden:
      description: Нет прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
  schemas:
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          examples: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT]
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    Organization:
      type: object
      required: [id, name, slug, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        planTier:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    OrganizationCreateRequest:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          maxLength: 200
        slug:
          type: string
          pattern: '^[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?$'
        planTier:
          type: string
          nullable: true
    OrganizationPatchRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
    Membership:
      type: object
      required: [id, userId, orgId, status, roles, joinedAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        status:
          type: string
          enum: [invited, active, suspended]
        roles:
          type: array
          items:
            type: string
        joinedAt:
          type: string
          format: date-time
          nullable: true
    MembershipPatchRequest:
      type: object
      properties:
        status:
          type: string
          enum: [invited, active, suspended]
        roleKeys:
          type: array
          items:
            type: string
          description: Полный набор ролей членства (замена)
    MemberPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Membership'
        nextCursor:
          type: string
          nullable: true
    Invite:
      type: object
      required: [id, email, expiresAt, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        roleKeys:
          type: array
          items:
            type: string
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        acceptedAt:
          type: string
          format: date-time
          nullable: true
    InviteCreateRequest:
      type: object
      required: [email, roleKeys]
      properties:
        email:
          type: string
          format: email
        roleKeys:
          type: array
          minItems: 1
          items:
            type: string
        ttlHours:
          type: integer
          minimum: 1
          maximum: 720
          default: 168
    InviteCreated:
      allOf:
        - $ref: '#/components/schemas/Invite'
        - type: object
          properties:
            acceptUrlHint:
              type: string
              description: Только для dev; в prod токен только в email
...
    OrgIdpConfigPublic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [oidc, saml]
        issuer:
          type: string
          nullable: true
        clientId:
          type: string
          nullable: true
        metadataUrl:
          type: string
          nullable: true
        enabled:
          type: boolean
        jitProvisioning:
          type: boolean
        defaultRoleKey:
          type: string
          nullable: true
    OrgIdpConfigPatchRequest:
      type: object
      properties:
        type:
          type: string
          enum: [oidc, saml]
        issuer:
          type: string
        clientId:
          type: string
        clientSecret:
          type: string
          description: Передаётся один раз; в ответе не возвращается
        metadataUrl:
          type: string
        enabled:
          type: boolean
        jitProvisioning:
          type: boolean
        defaultRoleKey:
          type: string
    OrgIdpTestRequest:
      type: object
      properties:
        metadataUrl:
          type: string
    OrgIdpTestResult:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        message:
          type: string
        issuerResolved:
          type: string
          nullable: true
...